feat: cleaned up publish-rc and extracted yarn utils
